pyproject.toml
[project]
name = "arrowlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "YEVGENYBURSHTEIN", email = "bursh@il.ibm.com" }
]
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.44.0",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"
